Extract github-related tests too
